youtube-summarize/requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
google-genai